aime_epic1/backend/fastapi/requirements.txt
fastapi>=0.112.0
uvicorn[standard]==0.30.1
pydantic>=2.8.2
python-dotenv>=1.0.1
httpx==0.27.0
requests==2.32.3
django==5.0.6
mysqlclient==2.2.4
whitenoise==6.6.0
gunicorn==22.0.0
langchain==0.2.14
langchain-community==0.2.12
ollama==0.3.1